Add script to run a single example
