+++ pyproject.toml
@@ -49,12 +49,18 @@ installer = "uv"
 
 [tool.hatch.envs.default.scripts]
 test-extension = "hatch test --extension {args}"
+
+# Build and serve documentation
 docs = [
     "hatch run docs:build",
     "hatch run docs:serve"
 ]
-example = "{env:HATCH_UV:uv} run scripts/run_examples/run_single_example.py {args}"
-all-examples = "{env:HATCH_UV:uv} run --isolated scripts/run_examples/run_all_examples.py {args:examples --zospy-location .}"
+
+# Run examples using the development version of ZOSPy
+example = "{env:HATCH_UV:uv} run --isolated scripts/run_examples/run_single_example.py {args} --zospy-location ."
+all-examples = "{env:HATCH_UV:uv} run --isolated scripts/run_examples/run_all_examples.py {args:examples} --zospy-location ."
+
+# Format docstrings according to numpydoc
 format-docstrings = "{env:HATCH_UV:uv} tool uvx pydocstringformatter --write {args:examples scripts tests zospy}"
 
 # Generate reference data for unit tests
